requirements.txt
# Core Web Framework
streamlit>=1.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Brokerage APIs
robin-stocks==3.0.5

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1

# Data Processing
pandas>=2.2.0
numpy>=2.0.0

# Visualization
plotly==5.18.0

# Market Data
yfinance>=0.2.48
tradingview-ta>=3.3.0
alpaca-trade-api>=3.0.2

# Web Scraping & Automation
selenium==4.16.0
beautifulsoup4==4.12.2
undetected-chromedriver==3.5.4

# HTTP Requests
requests==2.31.0
aiohttp==3.9.1
websocket-client==1.6.4

# Environment & Configuration
python-dotenv==1.0.0

# Authentication & Security
pyotp==2.9.0

# Scientific Computing
scipy==1.11.4
ta==0.11.0
pandas-ta>=0.4.71

# Options & Greeks
mibian==0.1.3

# Reddit API
praw==7.7.1

# Task Scheduling
celery==5.3.4
apscheduler==3.10.4

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0
psutil==5.9.6

# Notifications
python-telegram-bot==20.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain & LangGraph (Modern Agent Framework - 0.3+ for modern imports)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-groq>=0.2.0
langchain-google-genai>=2.0.0
langchain-huggingface>=0.1.0
langgraph>=0.2.50
langchain-text-splitters>=0.3.0

# Model Context Protocol (MCP)
mcp>=1.0.0

# AI/ML for Prediction Markets
sentence-transformers>=3.0.0
transformers>=4.40.0
tokenizers>=0.21,<0.22
torch>=2.0.0
scikit-learn>=1.4.0

# Vector Database & RAG
qdrant-client==1.7.0
chromadb==0.4.22

# LLM Providers (Updated to latest versions)
openai>=1.50.0
anthropic>=0.39.0
groq>=0.12.0
google-generativeai>=0.8.0
# deepseek==0.0.1  # Note: DeepSeek uses OpenAI-compatible API, no separate package needed

# Sports Data & Analytics
nfl-data-py>=0.3.0

# Odds & Betting APIs (Note: Install from pip if available)
# the-odds-api==1.0.0
